keep a buffer of undeliverable messages
In case the log server is not available, keep the messages
in a buffer, and try to send them later.

For this, add a common trait to the message structs, so they
can be stored in the same vector. Also, check before sending a
message is the socket is available. If it is not available,
try to connect to it. If the socket is still not available, then
store the message in the buffer. When the connection is available,
empty the buffer, and send the messages.