add convenience macro
The methods on loglib are nice, but their signature is not:
the formatter methods take a string (for the pattern) and and array of strings
(for the arguments) - which is definitely not a joy to use. However
variadic arguments aren't really a thing here.

Instead, here is a macro, that can take variadic arguments:
if it gets some arguments, it converts them to a vector, which is
passed to loglib as an array, so everyone is happy.